Split NetCore and DumpAgent into standalone crates
