+++ cmd/dump/Cargo.toml
@@ -10,8 +10,6 @@ clap.workspace = true
 goblin.workspace = true
 hif.workspace = true
 hubpack.workspace = true
-humility-cmd.workspace = true
-humility.workspace = true
 humpty.workspace = true
 indexmap.workspace = true
 indicatif.workspace = true
@@ -22,3 +20,7 @@ parse_int.workspace = true
 probe-rs.workspace = true
 zerocopy.workspace = true
 zip.workspace = true
+
+humility-dump-agent.workspace = true
+humility-cmd.workspace = true
+humility.workspace = true
