+++ humility-cmd/Cargo.toml
@@ -9,7 +9,6 @@ clap.workspace = true
 colored.workspace = true
 hif.workspace = true
 hubpack.workspace = true
-humility.workspace = true
 idol.workspace = true
 indexmap.workspace = true
 lazy_static.workspace = true
@@ -19,3 +18,6 @@ postcard.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 zerocopy.workspace = true
+
+humility.workspace = true
+humility-net-core.workspace = true
