+++ humility-core/Cargo.toml
@@ -19,12 +19,10 @@ idol.workspace = true
 indexmap.workspace = true
 indicatif.workspace = true
 log.workspace = true
-lzss.workspace = true
 multimap.workspace = true
 num-derive.workspace = true
 num-traits.workspace = true
 parse_int.workspace = true
-rand.workspace = true
 regex.workspace = true
 ron.workspace = true
 rusb.workspace = true
