+++ humility-core/src/core.rs
@@ -9,7 +9,6 @@ use anyhow::{anyhow, bail, ensure, Context, Result};
 
 use crate::arch::ARMRegister;
 use crate::hubris::*;
-use crate::net::decode_iface;
 use std::cell::RefCell;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
@@ -19,7 +18,6 @@ use std::fs;
 use std::io::Read;
 use std::io::Write;
 use std::net::TcpStream;
-use std::net::{ToSocketAddrs, UdpSocket};
 use std::path::Path;
 use std::rc::Rc;
 use std::str;
@@ -1373,276 +1371,6 @@ pub enum NetAgent {
     DumpAgent,
 }
 
-pub struct NetCore {
-    udprpc_socket: UdpSocket,
-    dump_agent_socket: UdpSocket,
-    flash: HubrisFlashMap,
-    ram: Vec<HubrisRegion>,
-}
-
-impl NetCore {
-    fn new(
-        ip: &str,
-        hubris: &HubrisArchive,
-        timeout: Duration,
-    ) -> Result<Self> {
-        let mut iter = ip.split('%');
-        let ip = iter.next().expect("ip address is empty");
-        let iface = iter
-            .next()
-            .ok_or_else(|| anyhow!("Missing scope id in IP (e.g. '%en0')"))?;
-
-        let scopeid = decode_iface(iface)?;
-
-        // See oxidecomputer/oana for standard Hubris UDP ports
-        let target = format!("[{}%{}]:998", ip, scopeid);
-
-        let dest = target.to_socket_addrs()?.collect::<Vec<_>>();
-        let udprpc_socket = UdpSocket::bind("[::]:0")?;
-        udprpc_socket.set_read_timeout(Some(timeout))?;
-        udprpc_socket.connect(&dest[..])?;
-
-        let rpc_task = hubris.lookup_task("udprpc").ok_or_else(|| {
-            anyhow!(
-                "Could not find `udprpc` task in this image. \
-                 Is it up to date?"
-            )
-        })?;
-        let _rpc_reply_type = hubris
-            .lookup_module(*rpc_task)?
-            .lookup_enum_byname(hubris, "RpcReply")?;
-
-        // See oxidecomputer/oana for standard Hubris UDP ports
-        let target = format!("[{}%{}]:11113", ip, scopeid);
-        let dest = target.to_socket_addrs()?.collect::<Vec<_>>();
-        let dump_agent_socket = UdpSocket::bind("[::]:0")?;
-        dump_agent_socket.set_read_timeout(Some(timeout))?;
-        dump_agent_socket.connect(&dest[..])?;
-
-        let mut out = Self {
-            udprpc_socket,
-            dump_agent_socket,
-            flash: HubrisFlashMap::new(hubris)?,
-            ram: vec![], // filled in below
-        };
-
-        // Surprisingly, this works: `hubris.regions` only reads data from the
-        // kernel's task table, which is in flash, so this works despite using a
-        // NetCore (which cannot read arbitrary memory).
-        let regions = hubris.regions(&mut out)?;
-        let mut ram_regions: Vec<_> = regions
-            .into_values()
-            .filter(|r| {
-                !r.attr.device
-                    && !r.attr.dma
-                    && !r.attr.execute
-                    && r.attr.read
-                    && r.attr.write
-                    && !r.tasks.is_empty()
-            })
-            .collect();
-
-        // Sort regions so we can find a target region with binary search
-        ram_regions.sort_by_key(|r| r.base);
-
-        // Assert a simple invariant to detect weird images
-        for (a, b) in ram_regions.iter().zip(ram_regions.iter().skip(1)) {
-            if b.base < a.base + a.size {
-                bail!("overlapping regions: {a:x?}; {b:x?}");
-            }
-        }
-
-        out.ram = ram_regions;
-        Ok(out)
-    }
-
-    fn read(&mut self, addr: u32, data: &mut [u8]) -> Result<()> {
-        if let Some((&base, &(size, offset))) =
-            self.flash.regions.range(..=addr).rev().next()
-        {
-            if base <= addr && base + size > addr {
-                let start = (addr - base) as usize;
-                let roffs = offset + start;
-
-                if start + data.len() <= size as usize {
-                    data.copy_from_slice(
-                        &self.flash.contents[roffs..roffs + data.len()],
-                    );
-
-                    return Ok(());
-                }
-
-                let len = (size as usize) - start;
-                data[..len]
-                    .copy_from_slice(&self.flash.contents[roffs..roffs + len]);
-
-                return self.read(addr + len as u32, &mut data[len..]);
-            }
-        }
-
-        self.read_ram(addr, data).context(format!(
-            "0x{addr:0x} can't be read via the archive or over the network"
-        ))
-    }
-
-    /// Reads data from RAM using the (UDP) dump agent as a proxy
-    fn read_ram(&mut self, addr: u32, data: &mut [u8]) -> Result<()> {
-        let p = self.ram.partition_point(|v| v.base <= addr);
-        if p == 0 {
-            bail!("address not found in RAM map");
-        }
-
-        // By construction, we know that every region in self.ram has at least
-        // one task associated with it.  If this task is the supervisor, then
-        // that's bad news.
-        let region = &self.ram[p - 1];
-        let Some(&task) = region.tasks.iter().find(|t| t.task() != 0) else {
-            bail!("supervisor memory cannot be read using dump facilities");
-        };
-
-        use crate::dump::{DumpAgent, DumpAgentExt, DumpArea, UdpDumpAgent};
-        let mut agent_core = crate::dump::AgentCore::new(self.flash.clone());
-
-        let mut udp_dump = UdpDumpAgent::new(self);
-
-        // Send a request to dump this specific region of memory within the task
-        let aligned_start = addr & !0b11;
-        let mut aligned_length = data.len() as u32 + addr - aligned_start;
-        while aligned_length & 0b11 != 0 {
-            aligned_length += 1;
-        }
-        let r = udp_dump.dump_remote_action(
-            humpty::udp::Request::DumpTaskRegion {
-                task_index: task.task(),
-                start: aligned_start,
-                length: aligned_length,
-            },
-        )?;
-        let dump_index = match r {
-            Ok(humpty::udp::Response::DumpTaskRegion(addr)) => addr,
-            Ok(o) => bail!("unexpected response from dump agent: {o:?}"),
-            Err(humpty::udp::Error::DumpAreaInUse) => {
-                bail!(
-                    "out of space for dump_agent; use \
-                     `humility dump --initialize-dump-agent` to free space"
-                )
-            }
-            Err(e) => {
-                bail!("dump agent failed: {e:?}")
-            }
-        };
-
-        // We're going to collect just this region into an ersatz AgentCore
-        (&mut udp_dump as &mut dyn DumpAgent).read_dump(
-            Some(DumpArea::ByIndex(dump_index as usize)),
-            &mut agent_core,
-            false,
-        )?;
-
-        // Pop the most recent dump, since we were just using it to read memory
-        // and it doesn't need to take up a dump area forever.
-        let r = udp_dump.dump_remote_action(
-            humpty::udp::Request::ReinitializeDumpFrom { index: dump_index },
-        )?;
-        match r {
-            Ok(humpty::udp::Response::ReinitializeDumpFrom) => (),
-            r => bail!("unexpected response from dump agent: {r:?}"),
-        }
-        // By construction, this AgentCore has exactly what it needs!
-        agent_core.read_8(addr, data)?;
-        Ok(())
-    }
-}
-
-#[rustfmt::skip::macros(bail)]
-impl Core for NetCore {
-    fn info(&self) -> (String, Option<String>) {
-        ("connected remotely".to_string(), None)
-    }
-
-    fn is_net(&self) -> bool {
-        true
-    }
-
-    fn set_timeout(&mut self, timeout: Duration) -> Result<()> {
-        self.udprpc_socket.set_read_timeout(Some(timeout))?;
-        self.dump_agent_socket.set_read_timeout(Some(timeout))?;
-        Ok(())
-    }
-
-    fn send(&self, buf: &[u8], target: NetAgent) -> Result<usize> {
-        match target {
-            NetAgent::UdpRpc => self.udprpc_socket.send(buf),
-            NetAgent::DumpAgent => self.dump_agent_socket.send(buf),
-        }
-        .map_err(anyhow::Error::from)
-    }
-
-    fn recv(&self, buf: &mut [u8], target: NetAgent) -> Result<usize> {
-        match target {
-            NetAgent::UdpRpc => self.udprpc_socket.recv(buf),
-            NetAgent::DumpAgent => self.dump_agent_socket.recv(buf),
-        }
-        .map_err(anyhow::Error::from)
-    }
-
-    fn read_8(&mut self, addr: u32, data: &mut [u8]) -> Result<()> {
-        self.read(addr, data)
-    }
-
-    fn read_reg(&mut self, reg: ARMRegister) -> Result<u32> {
-        bail!("cannot read register {} over network", reg);
-    }
-
-    fn write_reg(&mut self, reg: ARMRegister, _value: u32) -> Result<()> {
-        bail!("cannot write register {} over network", reg);
-    }
-
-    fn write_word_32(&mut self, _addr: u32, _data: u32) -> Result<()> {
-        bail!("cannot write a word over network");
-    }
-
-    fn write_8(&mut self, _addr: u32, _data: &[u8]) -> Result<()> {
-        bail!("cannot write a byte over network");
-    }
-
-    fn halt(&mut self) -> Result<()> {
-        Ok(())
-    }
-
-    fn run(&mut self) -> Result<()> {
-        Ok(())
-    }
-
-    fn step(&mut self) -> Result<()> {
-        bail!("can't step over network");
-    }
-
-    fn init_swv(&mut self) -> Result<()> {
-        bail!("cannot enable SWV over network");
-    }
-
-    fn read_swv(&mut self) -> Result<Vec<u8>> {
-        bail!("cannot read SWV over network");
-    }
-
-    fn load(&mut self, _path: &Path) -> Result<()> {
-        bail!("cannot load flash over network");
-    }
-
-    fn reset(&mut self) -> Result<()> {
-        bail!("cannot reset over network");
-    }
-
-    fn reset_and_halt(&mut self, _dur: std::time::Duration) -> Result<()> {
-        bail!("cannot reset over network");
-    }
-
-    fn wait_for_halt(&mut self, _dur: std::time::Duration) -> Result<()> {
-        bail!("cannot wait for halt over network");
-    }
-}
-
 fn parse_probe(probe: &str) -> (&str, Option<usize>) {
     if probe.contains('-') {
         let str = probe.to_owned();
@@ -1894,15 +1622,5 @@ pub fn attach_dump(
 ) -> Result<Box<dyn Core>> {
     let core = DumpCore::new(dump, hubris)?;
     crate::msg!("attached to dump");
-    Ok(Box::new(core))
-}
-
-pub fn attach_net(
-    ip: &str,
-    hubris: &HubrisArchive,
-    timeout: Duration,
-) -> Result<Box<dyn Core>> {
-    let core = NetCore::new(ip, hubris, timeout)?;
-    crate::msg!("connecting to {}", ip);
     Ok(Box::new(core))
 }
