+++ humility-core/src/lib.rs
@@ -11,7 +11,6 @@ use hubris::HubrisArchive;
 pub mod arch;
 pub mod cli;
 pub mod core;
-pub mod dump;
 pub mod env;
 pub mod hubris;
 pub mod net;
