+++ humility-net-core/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "humility-net-core"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+anyhow.workspace = true
+humpty.workspace = true
+
+humility.workspace = true
+humility-dump-agent.workspace = true
